[capstone] Find gadget roots based on an instruction set
Implements the find_gadget_roots function, that returns a vector of
offsets at which an instruction of interest is present in the given
code.

In this function,the capstone engine performs a first pass on the
given code, disassembling one isntruction at each bound-aligned offset
(using the ALIGNMENT constant).
If this instruction is present if the set of instructions of interest,
its offset is pushed to a vector returned by the function.

Signed-off-by: lfalkau <lfalkau@student.42.fr>