[core] Full rebuild of the core system
I choosed a radically different approach which will be documented soon.
In short, we first find all roots as before, but then we recursively try
to disassemble an instruction back until the limit instruction count is
reached or we encounter another jump instruction.

RaccoonV now has a linear complexity instead of a quadratic one.

Signed-off-by: Loic Buckwell <lfalkau@student.42.fr>